Make it clearer what `:host` does
I wasn't sure what the `:host` pseudo-selector was for until I dug through the closed issues.  Googling it brings up the shadow DOM and styling _it's_ root element but my brain was all "I'm not using the shadow DOM, maybe that's what this for? IDK".

Then I just played with it in the code and generated a page and realized it was exactly what I was looking for.  An example of the transformed CSS would go a long way to showing how that is uses.